Fix 5.11.2

+++ Assets/Scripts/AI/MovementProperties.cs
@@ -20,7 +20,7 @@ namespace BaseAI
         /// <summary>
         /// Количество дискретных углов поворота
         /// </summary>
-        public float angleSteps;
+        public int angleSteps;  
         /// <summary>
         /// Длина прыжка (фиксированная)
         /// </summary>
